+++ test/free_text_search_test.dart
@@ -11,7 +11,9 @@ import 'package:hive/hive.dart';
 import 'package:test/test.dart';
 import 'package:text_indexing/type_definitions.dart';
 import 'hashtag_analyzer.dart';
+import 'stock_news_index.dart';
 import 'test_utils.dart';
+import 'data/sample_news.dart';
 
 void main() {
   //
@@ -30,20 +32,19 @@ void main() {
       final service = await getService('hashtags');
       // final companyNames = await HashTagAnalyzer.getCompanyNames(service);
       final index = await HashTagIndex.hydrate();
-
-      // hiveIndex(() async => service.dataStore.length);
-      // Future<List<Token>> tokenFilter(List<Token> tokens) async {
-      //   final dft = await index.getDictionary(tokens.terms);
-      //   return tokens.where((token) => dft.keys.contains(token.term)).toList();
-      // }
-
       final documents = TestData.stockNews;
-      final documentZones = {'name': 1.0, 'description': 1.0};
+      final documentZones = {'name': 20.0, 'description': 1.0};
+      final weightingStrategy =
+          WeightingStrategy(zoneWeights: documentZones, positionThreshold: 0);
       for (final e in documents.entries) {
         final document = e.value;
         final name = document['name'];
-        final results = await FreeTextSearch(index)
-            .document(document, documentZones: documentZones, limit: 5);
+        final results = await FreeTextSearch(index).document(document,
+            weightingStrategy: weightingStrategy,
+            limit: 5,
+            nGramRange: NGramRange(1, 3),
+            tokenFilter: HashTagQueryAnalyzer.kFilterTokens,
+            docAnalyzer: HashTagQueryAnalyzer());
         // map the results for printing to console
         final JsonCollection jsonResults = {};
         for (final e in results) {
@@ -64,15 +65,15 @@ void main() {
 
     test('FreeTextSearch.startsWith', (() async {
       final phrase = 'nv';
-
+      final collection = sampleNews;
       // initialize an in-memory indexer
-      final indexer = await TestIndex.hydrate();
+      final index = await StockNewsIndex.hydrate();
       // get the results
-      final results = await FreeTextSearch(indexer.index).startsWith(phrase);
+      final results = await FreeTextSearch(index).startsWith(phrase);
       // map the results for printing to console
       final JsonCollection jsonResults = {};
       for (final e in results) {
-        final doc = indexer.collection[e.key];
+        final doc = collection[e.key];
         if (doc != null) {
           var name = (doc['name'] as String?) ?? '';
           name = name.length > 120 ? name.substring(0, 120) : name;
@@ -89,13 +90,14 @@ void main() {
     test('QuerySearch.search', () async {
       final phrase = 'dan ives wedbush';
       // initialize an in-memory indexer
-      final indexer = await TestIndex.hydrate();
+      // initialize an in-memory indexer
+      final index = await StockNewsIndex.hydrate();
       // get the results
-      final results = await FreeTextSearch(indexer.index).phrase(phrase);
+      final results = await FreeTextSearch(index).phrase(phrase);
 
       final JsonCollection jsonResults = {};
       for (final e in results) {
-        final doc = indexer.collection[e.docId];
+        final doc = StockNewsIndex.collection[e.docId];
         if (doc != null) {
           var name = (doc['name'] as String?) ?? '';
           name = name.length > 120 ? name.substring(0, 120) : name;